requirements.txt
requests
tinydb
python-dotenv
sqlalchemy
psycopg2
matplotlib
numpy
pandas
streamlit
plotly
